chore(601|602): aligned DTO for simple endpoints
- Aligned URL for code requests (org units and categories)
- Aligned DTOs for simple/code requests (org units, categories and forest client locations)
