+++ backend/src/main/java/ca/bc/gov/restapi/results/common/endpoint/ForestClientEndpoint.java
@@ -2,9 +2,9 @@ package ca.bc.gov.restapi.results.common.endpoint;
 
 import ca.bc.gov.restapi.results.common.dto.ForestClientAutocompleteResultDto;
 import ca.bc.gov.restapi.results.common.dto.ForestClientDto;
-import ca.bc.gov.restapi.results.common.dto.IdNameDto;
 import ca.bc.gov.restapi.results.common.exception.ForestClientNotFoundException;
 import ca.bc.gov.restapi.results.common.service.ForestClientService;
+import ca.bc.gov.restapi.results.oracle.dto.CodeDescriptionDto;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -47,7 +47,7 @@ public class ForestClientEndpoint {
   }
 
   @GetMapping("/{clientNumber}/locations")
-  public List<IdNameDto> getForestClientLocations(@PathVariable String clientNumber) {
+  public List<CodeDescriptionDto> getForestClientLocations(@PathVariable String clientNumber) {
     return forestClientService.getClientLocations(clientNumber);
   }
 }
