+++ backend/src/main/java/ca/bc/gov/restapi/results/oracle/service/OrgUnitService.java
@@ -1,6 +1,7 @@
 package ca.bc.gov.restapi.results.oracle.service;
 
 import ca.bc.gov.restapi.results.common.configuration.SilvaConfiguration;
+import ca.bc.gov.restapi.results.oracle.dto.CodeDescriptionDto;
 import ca.bc.gov.restapi.results.oracle.entity.OrgUnitEntity;
 import ca.bc.gov.restapi.results.oracle.repository.OrgUnitRepository;
 import java.util.List;
@@ -25,7 +26,7 @@ public class OrgUnitService {
    *
    * @return List of {@link OrgUnitEntity} with found categories.
    */
-  public List<OrgUnitEntity> findAllOrgUnits() {
+  public List<CodeDescriptionDto> findAllOrgUnits() {
     log.info("Getting all org units for the search openings");
 
     if (Objects.isNull(silvaConfiguration.getOrgUnits())
@@ -35,8 +36,11 @@ public class OrgUnitService {
       return List.of();
     }
 
-    List<OrgUnitEntity> orgUnits = orgUnitRepository.findAllByOrgUnitCodeIn(
-        silvaConfiguration.getOrgUnits());
+    List<CodeDescriptionDto> orgUnits = orgUnitRepository
+        .findAllByOrgUnitCodeIn(silvaConfiguration.getOrgUnits())
+        .stream()
+        .map(orgUnit -> new CodeDescriptionDto(orgUnit.getOrgUnitCode(), orgUnit.getOrgUnitName()))
+        .toList();
 
     log.info("Found {} org units by codes", orgUnits.size());
     return orgUnits;
