+++ frontend/src/services/OpeningService.ts
@@ -1,17 +1,13 @@
 import axios from 'axios';
 import { getAuthIdToken } from './AuthService';
-import { env } from '../env';
 import {
   IOpeningPerYear,
-  IFreeGrowingProps,
-  IFreeGrowingChartData,
   PaginatedRecentOpeningsDto,
-  OrgUnitEntity,
   OpeningsPerYearDto
 } from '../types/OpeningTypes';
 import { API_ENDPOINTS, defaultHeaders } from '@/services/apiConfig';
+import CodeDescriptionDto from '@/types/CodeDescriptionType';
 
-const backendUrl = env.VITE_BACKEND_URL;
 
 /**
  * Fetch users submission trends.
@@ -53,54 +49,6 @@ export async function fetchUserSubmissionTrends(props: IOpeningPerYear): Promise
   }
 }
 
-/**
- * Fetch free growing milestones data from backend.
- *
- * @returns {IFreeGrowingChartData[]} Array with recent action objects.
- */
-export async function fetchFreeGrowingMilestones(props: IFreeGrowingProps): Promise<IFreeGrowingChartData[]> {
-  const authToken = getAuthIdToken();
-  let url = backendUrl.concat("/api/dashboard-metrics/free-growing-milestones");
-
-  // Construct URL with optional parameters
-  if (props.orgUnitCode || props.clientNumber || props.entryDateStart || props.entryDateEnd) {
-    url += '?';
-    if (props.orgUnitCode) url += `orgUnitCode=${props.orgUnitCode}&`;
-    if (props.clientNumber) url += `clientNumber=${props.clientNumber}&`;
-    if (props.entryDateStart) url += `entryDateStart=${props.entryDateStart}&`;
-    if (props.entryDateEnd) url += `entryDateEnd=${props.entryDateEnd}&`;
-    // Remove trailing '&' if present
-    url = url.replace(/&$/, '');
-  }
-
-  try {
-    const response = await axios.get(url, {
-      headers: {
-        'Content-Type': 'application/json',
-        'Access-Control-Allow-Origin': window.location.origin,
-        Authorization: `Bearer ${authToken}`
-      }
-    });
-
-    const { data } = response;
-    if (data && Array.isArray(data)) {
-      // Transform data for DonutChartView component
-      const transformedData: IFreeGrowingChartData[] = data.map(item => ({
-        group: item.label,
-        value: item.amount
-      }));
-
-      return transformedData;
-    } else {
-      return [];
-    }
-  } catch (error) {
-    console.error('Error fetching free growing milestones:', error);
-    throw error;
-  }
-}
-
-
 // Used to fetch the recent openings for a user
 export const fetchUserRecentOpenings = (): Promise<PaginatedRecentOpeningsDto> => {
 
@@ -115,7 +63,7 @@ export const fetchUserRecentOpenings = (): Promise<PaginatedRecentOpeningsDto> =
 /**
  * Fetch a list of org unit used for opening search
  */
-export const fetchOpeningsOrgUnits = (): Promise<OrgUnitEntity[]> => {
+export const fetchOpeningsOrgUnits = (): Promise<CodeDescriptionDto[]> => {
   const authToken = getAuthIdToken();
 
   return axios.get(API_ENDPOINTS.orgUnits(),defaultHeaders(authToken))
